backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.2.1",
    "fastapi[standard]>=0.135.3",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.49",
]
